Move firebase-vertex to firebase-vertexai
