+++ firebase-vertexai/gradle.properties
