+++ firebase-vertexai/src/main/kotlin/com/google/firebase/vertexai/type/UsageMetadata.kt
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2024 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.firebase.vertexai.type
+
+/**
+ * Usage metadata about response(s).
+ *
+ * @param promptTokenCount Number of tokens in the request.
+ * @param candidatesTokenCount Number of tokens in the response(s).
+ * @param totalTokenCount Total number of tokens.
+ */
+class UsageMetadata(
+  val promptTokenCount: Int,
+  val candidatesTokenCount: Int,
+  val totalTokenCount: Int
+)
