+++ subprojects.cfg
@@ -57,7 +57,7 @@ firebase-sessions
 firebase-sessions:test-app
 firebase-storage
 firebase-storage:ktx
-firebase-vertex
+firebase-vertexai
 protolite-well-known-types
 
 encoders
